added saveWIshList and saveReadList button functionalities

+++ src/components/localStorage.js
@@ -17,21 +17,37 @@ export const getWishListBooks = () => {
 
 export const saveReadBook = (book) => {
   let books = getBooks();
-//   const isExist = books.find((blg) => blg.bookId == book.bookId);
-  //   if (isExist) {
-  //     return toast.error("Already Bookmarked!");
-  //   }
+  const isExist = books.find((blg) => blg.bookId == book.bookId);
+  if (isExist) {
+    return alert("Already read the book");
+  }
   books.push(book);
   localStorage.setItem("books", JSON.stringify(books));
-  //   toast.success("Bookmarked Successfully");
+  alert("added in readlist Successfully");
 };
-export const saveWishListdBook = (book) => {
-  let books = getBooks();
-//   const isExist = books.find((blg) => blg.bookId == book.bookId);
-  //   if (isExist) {
-  //     return toast.error("Already Bookmarked!");
-  //   }
+
+export const saveWishListdBook = (book, id) => {
+  let books = getWishListBooks();
+  const readListBooks = getBooks();
+  // console.log(readListBooks);
+
+  // Check if the book is already in the wishlist
+  const isExistBook = books.find((blg) => blg.bookId == id);
+
+  const isExistReadListBook = readListBooks.find(
+    (readListBook) => readListBook.bookId == id
+  );
+  if (isExistReadListBook) {
+    // Check if the book is also in the read list
+    return console.log("Already read the book");
+  }
+  if (isExistBook) {
+    // Book is in the read list, alert and return
+    return console.log("Already added in the wishList");
+  }
+
+  // Book not in wishlist or read list, add to wishlist
   books.push(book);
   localStorage.setItem("wish-books", JSON.stringify(books));
-  //   toast.success("Bookmarked Successfully");
+  return console.log("Successfully added in wishlists");
 };
